Enhancements

+++ service.sh
@@ -16,6 +16,51 @@ resetprop audio.dolby.ds2.hardbypass true
 resetprop vendor.audio.dolby.ds2.enabled true
 resetprop vendor.audio.dolby.ds2.hardbypass true
 
+# function
+stop_service() {
+for NAMES in $NAME; do
+  if getprop | grep "init.svc.$NAMES\]: \[running"; then
+    stop $NAMES
+  fi
+done
+}
+run_service() {
+for FILES in $FILE; do
+  killall $FILES
+  $FILES &
+  PID=`pidof $FILES`
+done
+}
+
+# stop
+NAME="dms-hal-1-0 dms-hal-2-0 dms-v36-hal-2-0"
+stop_service
+
+# mount
+DIR=/odm/bin/hw
+FILE=$DIR/vendor.dolby_v3_6.hardware.dms360@2.0-service
+if [ "`realpath $DIR`" == $DIR ] && [ -f $FILE ]; then
+  mount -o bind $MODPATH/system/vendor/$FILE $FILE
+fi
+
+# run
+FILE=`realpath /vendor`/bin/hw/vendor.dolby.hardware.dms@1.0-service
+run_service
+
+# restart
+VIBRATOR=`realpath /*/bin/hw/vendor.qti.hardware.vibrator.service*`
+[ "$VIBRATOR" ] && killall $VIBRATOR
+POWER=`realpath /*/bin/hw/vendor.mediatek.hardware.mtkpower@*-service`
+[ "$POWER" ] && killall $POWER
+killall android.hardware.usb@1.0-service
+killall android.hardware.usb@1.0-service.basic
+killall android.hardware.sensors@1.0-service
+killall android.hardware.sensors@2.0-service-mediatek
+killall android.hardware.light-service.mt6768
+killall android.hardware.lights-service.xiaomi_mithorium
+CAMERA=`realpath /*/bin/hw/android.hardware.camera.provider@*-service_64`
+[ "$CAMERA" ] && killall $CAMERA
+
 # wait
 sleep 20
 
@@ -54,58 +99,23 @@ if [ -d /my_product/etc ] && [ "$FILE" ]; then
 fi
 
 # restart
-killall audioserver
-
-# function
-stop_service() {
-for NAMES in $NAME; do
-  if getprop | grep "init.svc.$NAMES\]: \[running"; then
-    stop $NAMES
-  fi
-done
-}
-run_service() {
-for FILES in $FILE; do
-  killall $FILES
-  $FILES &
-  PID=`pidof $FILES`
-done
-}
-
-# stop
-NAME="dms-hal-1-0 dms-hal-2-0 dms-v36-hal-2-0"
-stop_service
-
-# run
-FILE=`realpath /vendor`/bin/hw/vendor.dolby.hardware.dms@1.0-service
-run_service
-
-# restart
-killall com.dolby.daxservice
-VIBRATOR=`realpath /*/bin/hw/vendor.qti.hardware.vibrator.service*`
-[ "$VIBRATOR" ] && killall $VIBRATOR
-POWER=`realpath /*/bin/hw/vendor.mediatek.hardware.mtkpower@*-service`
-[ "$POWER" ] && killall $POWER
-killall android.hardware.usb@1.0-service
-killall android.hardware.usb@1.0-service.basic
-killall android.hardware.sensors@1.0-service
-killall android.hardware.sensors@2.0-service-mediatek
-killall android.hardware.light-service.mt6768
-killall android.hardware.lights-service.xiaomi_mithorium
-CAMERA=`realpath /*/bin/hw/android.hardware.camera.provider@*-service_64`
-[ "$CAMERA" ] && killall $CAMERA
+PID=`pidof audioserver`
+if [ "$PID" ]; then
+  killall audioserver
+fi
 
 # wait
 sleep 40
 
 # allow
-PKG=com.dolby.dax2appUI
+PKG=com.dolby.daxservice
 if [ "$API" -ge 30 ]; then
   appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
 fi
+killall $PKG
 
 # allow
-PKG=com.dolby.daxservice
+PKG=com.dolby.dax2appUI
 if [ "$API" -ge 30 ]; then
   appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
 fi
